+++ diracx-db/src/diracx/db/sql/__init__.py
@@ -3,5 +3,7 @@ from __future__ import annotations
 __all__ = ("AuthDB", "JobDB", "JobLoggingDB", "SandboxMetadataDB", "TaskQueueDB")
 
 from .auth.db import AuthDB
-from .jobs.db import JobDB, JobLoggingDB, TaskQueueDB
+from .job.db import JobDB
+from .job_logging.db import JobLoggingDB
 from .sandbox_metadata.db import SandboxMetadataDB
+from .task_queue.db import TaskQueueDB
